@@ -342,50 +344,47 @@
     "css-hot-loader": "^1.4.4",
     "css-loader": "^2.1.1",
     "enzyme": "^3.9.0",
-    "enzyme-adapter-react-16": "^1.12.1",
+    "enzyme-adapter-react-16": "^1.13.2",
     "eslint": "^5.3.0",
     "eslint-config-airbnb": "^17.1.0",
-    "eslint-config-prettier": "^4.2.0",
+    "eslint-config-prettier": "^4.3.0",
     "eslint-import-resolver-webpack": "^0.11.1",
     "eslint-plugin-import": "^2.14.0",
     "eslint-plugin-jsx-a11y": "^6.1.1",
-    "eslint-plugin-prettier": "^3.0.1",
+    "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-react": "^7.11.0",
     "file-loader": "^3.0.1",
-    "flow-bin": "^0.98.0",
+    "flow-bin": "^0.98.1",
     "friendly-errors-webpack-plugin": "^1.7.0",
-    "husky": "^2.2.0",
+    "husky": "^2.3.0",
     "identity-obj-proxy": "^3.0.0",
     "imagemin-webpack-plugin": "^2.4.2",
-    "jest": "^24.7.1",
-    "lint-staged": "^8.1.5",
+    "jest": "^24.8.0",
+    "lint-staged": "^8.1.7",
     "mini-css-extract-plugin": "^0.6.0",
     "nock": "^10.0.6",
-    "nodemon": "^1.19.0",
+    "nodemon": "^1.19.1",
     "npm-run-all": "^4.1.5",
     "optimize-css-assets-webpack-plugin": "^5.0.1",
-    "postcss": "^7.0.14",
+    "postcss": "^7.0.16",
     "postcss-loader": "^3.0.0",
-    "prettier": "^1.17.0",
+    "prettier": "^1.17.1",
     "raf": "^3.4.1",
     "react-test-renderer": "^16.8.6",
     "redux-mock-store": "^1.5.3",
     "rimraf": "^2.6.3",
     "sass-loader": "^7.1.0",
     "stylelint": "^10.0.1",
-    "stylelint-config-prettier": "^5.1.0",
+    "stylelint-config-prettier": "^5.2.0",
     "stylelint-config-recommended-scss": "^3.3.0",
     "stylelint-config-standard": "^18.3.0",
     "stylelint-scss": "^3.6.1",
     "url-loader": "^1.1.2",
-    "webpack": "^4.30.0",
+    "webpack": "^4.32.2",
     "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.3.1",
-    "webpack-dev-middleware": "^3.6.2",
-    "webpack-hot-middleware": "^2.24.4",
+    "webpack-cli": "^3.3.2",
+    "webpack-dev-middleware": "^3.7.0",
+    "webpack-hot-middleware": "^2.25.0",
     "webpack-manifest-plugin": "^2.0.4"
-  },
-  "resolutions": {
-    "webpack-sources": "^1.2.0"
   }
 }
