@@ -1,8 +1,8 @@
 /* @flow */
 
-module.exports = {
+export default {
   host: process.env.NODE_HOST || 'localhost', // Define your host from 'package.json'
-  port: process.env.PORT,
+  port: process.env.PORT || '8080',
   app: {
     htmlAttributes: { lang: 'en' },
     title: 'React Cool Starter',
