@@ -2,15 +2,14 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { MemoryRouter } from 'react-router-dom';
 
-import ErrorDisplay from '../index';
+import ErrorBoundary from '../index';
 
-describe('<ErrorDisplay />', () => {
+describe('<ErrorBoundary />', () => {
   it('renders', () => {
-    const mockData = { message: 'Error!' };
     const tree = renderer
       .create(
         <MemoryRouter>
-          <ErrorDisplay error={mockData} />
+          <ErrorBoundary />
         </MemoryRouter>
       )
       .toJSON();
